Remove legacy AZC0004 analyzer
