+++ src/dotnet/Azure.ClientSdk.Analyzers/Azure.ClientSdk.Analyzers/Descriptors.cs
@@ -24,14 +24,6 @@ namespace Azure.ClientSdk.Analyzers
             "https://azure.github.io/azure-sdk/dotnet_introduction.html#dotnet-service-methods-virtual"
         );
 
-        public static DiagnosticDescriptor AZC0004 = new DiagnosticDescriptor(
-            nameof(AZC0004),
-            "DO provide both asynchronous and synchronous variants for all service methods.",
-            "DO provide both asynchronous and synchronous variants for all service methods.",
-            DiagnosticCategory.Usage, DiagnosticSeverity.Warning, isEnabledByDefault: true, description: null,
-            "https://azure.github.io/azure-sdk/dotnet_introduction.html#dotnet-service-methods-sync-and-async"
-        );
-
         public static DiagnosticDescriptor AZC0005 = new DiagnosticDescriptor(
             nameof(AZC0005),
             "DO provide protected parameterless constructor for mocking.",
